Add files via upload
Project description
This project “The Travel Agency Management System in C Programming” is complete and totally-error free clean code. The main goal of this project is to Book and manage the data of booking of vehicles from agency.
Basically this project console based project in c Programming Language so here’s no use of the graphics.
The works which will be performed by the system are:
*Whole Project is password protected (First screen is Login Window having user id and password)
*Note - Here user is owner of travel agency
•	User can add the Vehicles (Car/four wheelers).
•	User can see the details of vehicles which was added by the user.
•	User can also delete a particular vehicle/car if there is no longer use of it.
•	After adding, editing and deleting the main function that our system will perform that is Booking Vehicles/Car for customer.
•	In this user can book a car/vehicle for customer on a rent for a period of time.
•	After Booking or rent a vehicle user can check the Booking Details.
•	In Booking details user can see the booked car , customer name Pickup and Drop point and also the date of the Journey
•	After this user can also check the updated vehicle list/available vehicle list after the booking.
•	And Command function is for exit the system. 
Here are the functions used in my project:
1.	addVehicleDetails() -for adding the vehicles
2.	showVehicleDetails()- for showing the vehicles which was added.
3.	DeleteVehicle()-for deleting vehicle
4.	bookVehicle()-for booking(rent) vehicle for customer
5.	bookedVehicleDetails()-for seeing the details of booked vehicles.
6.	availableVehicleDetails()-for seeing updated vehicle list.

Data Storing in the memory:
I am using “File Handling” for storing or saving the data in the memory 
File handling in C refers to the task of storing data in the form of input or output produced by running C programs in data files, namely, a text file or a binary file for future reference and analysis.
So far, the operations using the C program are done on a prompt / terminal which is not stored anywhere. But in the software industry, most of the programs are written to store the information fetched from the program.